Final Dashboard added
Dataset Description

Dataset is about a company's sales made from multiple departmental stores across the world, with details of the sale, discount, profit, product, customer, store, and location data.

Business problem

To analyze and find the regions and product categories for improvement in business, also locate the best performing products and our top customers on the way. To find our Delivery performance and department level analysis on sales and profits.